+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Changed that resource link contains uri and the current request will be used for resource expanding. 
 - Changed middleware order for HATEOAS middleware to have a identified request.
 - Changed optional object property validator from array with added boolean value to ProxyValidator.
+- Changed BetweenValidator internal value validator from numeric to integer.
 
 ### Deprecated